Merge dev to main for final delivery

See merge request idatt1002-2022-k2-02/idatt1002_2022_k2_02!112

+++ quizmaker/src/main/java/org/ntnu/k2/g2/quizmaker/data/TeamModel.java
@@ -7,40 +7,43 @@ import java.util.Objects;
  */
 public class TeamModel {
     private String teamName;
-    private int score;
+    private double score;
     private int id = -1;
 
     /**
-     * Constructs a new Team-instance. This method is only used by the QuizRegister.
-     * New teams should be created using the QuizRegister::newTeam method.
+     * Constructs a new Team-instance. This method is only used by the QuizRegister. New teams should be created using
+     * the QuizRegister::newTeam method.
      */
-    protected TeamModel(){}
+    protected TeamModel() {
+    }
 
     @Override
     public String toString() {
-        return "Team{" +
-                "id=" + id +
-                ", teamName='" + teamName + '\'' +
-                ", score=" + score +
-                '}';
+        return "Team{" + "id=" + id + ", teamName='" + teamName + '\'' + ", score=" + score + '}';
     }
 
     /**
      * Checks the equality of two teams. All parameters are considered.
-     * @param o The object to compare to.
-     * @return The equality of the two teams. If the two teams an all properties are equal, true is returned.
-     *         Otherwise, false is returned.
+     *
+     * @param o
+     *            The object to compare to.
+     *
+     * @return The equality of the two teams. If the two teams an all properties are equal, true is returned. Otherwise,
+     *         false is returned.
      */
     @Override
     public boolean equals(Object o) {
-        if (this == o) return true;
-        if (o == null || getClass() != o.getClass()) return false;
+        if (this == o)
+            return true;
+        if (o == null || getClass() != o.getClass())
+            return false;
         TeamModel team = (TeamModel) o;
         return score == team.score && id == team.id && Objects.equals(teamName, team.teamName);
     }
 
     /**
      * The HashCode of the object. All properties are considered.
+     *
      * @return The HashCode of the object using Objects.hash()
      */
     @Override
@@ -57,7 +60,9 @@ public class TeamModel {
 
     /**
      * Sets the id of the team. This is done at creation by TeamDAO.
-     * @param id The id of the team.
+     *
+     * @param id
+     *            The id of the team.
      */
     protected void setId(int id) {
         this.id = id;
@@ -71,9 +76,11 @@ public class TeamModel {
     }
 
     /**
-     * Sets the name of the team. Changes to this property will be reflected in the database
-     * after QuizRegister::saveTeam is called.
-     * @param teamName The name of the team.
+     * Sets the name of the team. Changes to this property will be reflected in the database after
+     * QuizRegister::saveTeam is called.
+     *
+     * @param teamName
+     *            The name of the team.
      */
     public void setTeamName(String teamName) {
         this.teamName = teamName;
@@ -82,16 +89,18 @@ public class TeamModel {
     /**
      * @return The score of the quiz.
      */
-    public int getScore() {
+    public double getScore() {
         return score;
     }
 
     /**
-     * Sets the score of the team. Changes to this property will be reflected in the database
-     * after QuizRegister::saveTeam is called.
-     * @param score The score of the team.
+     * Sets the score of the team. Changes to this property will be reflected in the database after
+     * QuizRegister::saveTeam is called.
+     *
+     * @param score
+     *            The score of the team.
      */
-    public void setScore(int score) {
+    public void setScore(double score) {
         this.score = score;
     }
 }
